Refactor: BaseBot separated from StaffBot
