+++ orchestra/communication/urls.py
@@ -6,8 +6,8 @@ from orchestra.communication.views import reject_staffing_request
 
 urlpatterns = [
     # Interface API
-    url(r'^accept_request/(?P<pk>[0-9]+)/$',
+    url(r'^accept_request/(?P<staffing_request_id>[0-9]+)/$',
         accept_staffing_request, name='accept_staffing_request'),
-    url(r'^reject_request/(?P<pk>[0-9]+)/$',
+    url(r'^reject_request/(?P<staffing_request_id>[0-9]+)/$',
         reject_staffing_request, name='reject_staffing_request'),
 ]
